+++ R/storage_utils.R
@@ -244,3 +244,19 @@ delete_confirmed <- function(confirm, name, type)
     isTRUE(ok)
 }
 
+
+render_xml <- function(lst)
+{
+    xml <- xml2::as_xml_document(lst)
+    rc <- rawConnection(raw(0), "w")
+    on.exit(close(rc))
+    xml2::write_xml(xml, rc)
+    rawToChar(rawConnectionValue(rc))
+}
+
+
+sign_sha256 <- function(string, key)
+{
+    openssl::base64_encode(openssl::sha256(charToRaw(string), openssl::base64_decode(key)))
+}
+