+++ tests/testthat/setup.R
@@ -1,0 +1,4 @@
+make_name <- function(n=20)
+{
+    paste0(sample(letters, n, TRUE), collapse="")
+}